MvC2: shuma fix, beginning of patch functionality
Shuma's Extras had an incorrect step and so were corrupting Spider-man.  Fixed that and also handled a Shuma stance we weren't handling.

This also has most of a new "Patch" functionality for previously created edits.  Still need to handle Spiral, wire it up to UI correctly, etc.

Also fixed some UINT8/UINT16 and other code sanity issues.

+++ palmod/Game/Game_SFA3_A.cpp
@@ -318,7 +318,7 @@ void CGame_SFA3_A::UpdatePalData()
 			int nIndexStart = 1;
 
 			COLORREF * crSrc = srcDef->pPal;
-			UINT8 uAmt = srcDef->uPalSz;
+			UINT16 uAmt = srcDef->uPalSz;
 
 			for(int nPICtr = nIndexStart; nPICtr < uAmt; nPICtr++)
 			{
